+++ src/graphics/graphicsproteinitem.cpp
@@ -130,7 +130,7 @@ void GraphicsProteinItem::setPolymer(const Polymer *polymer)
             QList<AminoAcid *> residues;
             AminoAcid::Conformation conformation = static_cast<AminoAcid *>(chain->residue(0))->conformation();
 
-            for(int i = 0; i < chain->size(); i++){
+            for(size_t i = 0; i < chain->size(); i++){
                 AminoAcid *residue = static_cast<AminoAcid *>(chain->residue(i));
 
                 if(residue->conformation() != conformation){
